Moved DataDirConstants to data_utils
